Made Shell for Crypt Level
Made main part of the shell for the crypt level I walked through as the player the scale seemed good in my opinion I can fix it if it seems too big or too small. I also made prefabs for the columns to prevent any problems since they are circular.