Sewer Level add assets pass 2
Notes:
Many assets appear to be missing, I.E.: The rat enemy, the fuse, the ladders.
These will need to be added in and properly set up to function in their intended ways.

Changes made:
Torches, gates, and key items have been added with textures and the like.

TODO:
Add in remaining missing assets/replace placeholder items with proper assets.
Fix lighting issues - torches do not emit light properly and the walls of parts of the sewer are still too shiny (simply turning down the metallic slider does not fix this)
Compile all parts of current sewer levels into one and test AI pathing, tracking, attacking, etc.
Test if box colliders could be used for the shell collision (due to crawlspaces and geometry with the pipes, this may not be possible)