Added Torches to Crypt and JumpScares to Sewer
Added a jumpscare at the beginning of the  sewer level when enering the crawl space, lets the player know there are enemies in the sewer. Also added some audio ones that can be used one when picking up a key item and one when getting through the final gate where it plays church bells as an indicator of the next level.